refactor: modularized edit audios on case and homologued styles
